requirements.txt
Django==1.5.1
cython
numpy==1.7.1
scipy
matplotlib==1.2.1
django-tastypie
simplejson==3.3.0
### Development ###
south
mock
mysql
mysql-python
selenium
factory_boy
